Refactor package imports
