+++ data/maga.txt
@@ -1,2 +1,1 @@
 T | 0 | 1
-T | 0 | murder the dogs and murder the pets
